Move Cobalt DB import to its own file
